da-model/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pjm-da-forecast"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=12.0.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "shap>=0.43.0",
    "statsmodels>=0.14.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "joblib>=1.3.0",
    "pytz",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
